Reproduce classpath according to export-dep-as-jar
Previously, fastpass didn't interpret the output of export-dep-as-jar
correctly resulting in both slower compilation (too many jars on the
compile classpath) and compile errors (missing jars on the classpath).
This commit changes fastpass to make it use exactly the classpath
that is produced by export-dep-as-jar.

This change is a regression at the moment because export-dep-as-jar
still misses some jars from the classpath. We'll need to fix that issue
in Pants first before moving ahead with this change.